@@ -21,6 +21,7 @@ using Altinn.Platform.Authentication.Services.Interfaces;
 using Altinn.Platform.Profile.Models;
 
 using AltinnCore.Authentication.Constants;
+
 using Microsoft.AspNetCore.Antiforgery;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
@@ -127,9 +128,9 @@ namespace Altinn.Platform.Authentication.Controllers
         [HttpGet("authentication")]
         public async Task<ActionResult> AuthenticateUser([FromQuery] string goTo, [FromQuery] bool dontChooseReportee)
         {
-            if (string.IsNullOrEmpty(goTo) && HttpContext.Request.Cookies[_generalSettings.AuthnGotToCookieName] != null)
+            if (string.IsNullOrEmpty(goTo) && HttpContext.Request.Cookies[_generalSettings.AuthnGoToCookieName] != null)
             {
-                goTo = HttpContext.Request.Cookies[_generalSettings.AuthnGotToCookieName];
+                goTo = HttpContext.Request.Cookies[_generalSettings.AuthnGoToCookieName];
             }
 
             if (!Uri.TryCreate(goTo, UriKind.Absolute, out Uri goToUri) || !IsValidRedirectUri(goToUri.Host))
@@ -215,7 +216,7 @@ namespace Altinn.Platform.Authentication.Controllers
                 }
                 catch (SblBridgeResponseException sblBridgeException)
                 {
-                    _logger.LogWarning($"SBL Bridge replied with {sblBridgeException.Response.StatusCode} - {sblBridgeException.Response.ReasonPhrase}");
+                    _logger.LogWarning(sblBridgeException, "SBL Bridge replied with {StatusCode} - {ReasonPhrase}", sblBridgeException.Response.StatusCode, sblBridgeException.Response.ReasonPhrase);
                     return StatusCode(StatusCodes.Status503ServiceUnavailable);
                 }
             }
@@ -350,7 +351,7 @@ namespace Altinn.Platform.Authentication.Controllers
             }
             catch (Exception ex)
             {
-                _logger.LogWarning($"Altinn Studio authentication failed. {ex.Message}");
+                _logger.LogWarning(ex, "Altinn Studio authentication failed.");
                 return Unauthorized();
             }
         }
@@ -383,7 +384,7 @@ namespace Altinn.Platform.Authentication.Controllers
                 string issOriginal = originalPrincipal.Claims.Where(c => c.Type.Equals(IssClaimName)).Select(c => c.Value).FirstOrDefault();
                 if (issOriginal == null || !_generalSettings.MaskinportenWellKnownConfigEndpoint.Contains(issOriginal))
                 {
-                    _logger.LogInformation("Invalid issuer " + issOriginal);
+                    _logger.LogInformation("Invalid issuer {issOriginal}", issOriginal);
                     return Unauthorized();
                 }
 
@@ -469,12 +470,12 @@ namespace Altinn.Platform.Authentication.Controllers
             }
             catch (Exception ex)
             {
-                _logger.LogWarning($"Organisation authentication failed. {ex.Message}");
+                _logger.LogWarning(ex, "Organisation authentication failed.");
                 return Unauthorized();
             }
         }
 
-        private async Task<(UserAuthenticationResult, ActionResult)> HandleEnterpriseUserLogin(string enterpriseUserHeader, string orgNumber)
+        private async Task<(UserAuthenticationResult AuthenticatedEnterpriseUser, ActionResult Error)> HandleEnterpriseUserLogin(string enterpriseUserHeader, string orgNumber)
         {
             EnterpriseUserCredentials credentials;
 
@@ -583,7 +584,7 @@ namespace Altinn.Platform.Authentication.Controllers
             }
             catch (Exception ex)
             {
-                _logger.LogWarning($"End user system authentication failed. {ex.Message}");
+                _logger.LogWarning(ex, "End user system authentication failed.");
                 return Unauthorized();
             }
         }
@@ -731,8 +732,8 @@ namespace Altinn.Platform.Authentication.Controllers
             UserAuthenticationModel userAuthenticationModel = new UserAuthenticationModel()
             {
                 IsAuthenticated = true,
-                ProviderClaims = new Dictionary<string,
-                List<string>>(), Iss = provider.IssuerKey,
+                ProviderClaims = new Dictionary<string, List<string>>(),
+                Iss = provider.IssuerKey,
                 AuthenticationMethod = AuthenticationMethod.NotDefined
             };
 
@@ -802,7 +803,7 @@ namespace Altinn.Platform.Authentication.Controllers
                     {
                         userAuthenticationModel.ProviderClaims.Add(claim.Type, new List<string>());
                     }
-                       
+
                     userAuthenticationModel.ProviderClaims[claim.Type].Add(claim.Value);
                 }
             }
@@ -817,7 +818,7 @@ namespace Altinn.Platform.Authentication.Controllers
 
         private async Task IdentifyOrCreateAltinnUser(UserAuthenticationModel userAuthenticationModel, OidcProvider provider)
         {
-            UserProfile profile = null;
+            UserProfile profile;
 
             if (!string.IsNullOrEmpty(userAuthenticationModel.ExternalIdentity))
             {
@@ -831,15 +832,17 @@ namespace Altinn.Platform.Authentication.Controllers
                     return;
                 }
 
-                UserProfile userToCreate = new UserProfile();
-                userToCreate.ExternalIdentity = issExternalIdentity;
-                userToCreate.UserName = CreateUserName(userAuthenticationModel, provider);
-                userToCreate.UserType = Profile.Enums.UserType.SelfIdentified;
+                UserProfile userToCreate = new()
+                {
+                    ExternalIdentity = issExternalIdentity,
+                    UserName = CreateUserName(userAuthenticationModel, provider),
+                    UserType = Profile.Enums.UserType.SelfIdentified
+                };
 
                 UserProfile userCreated = await _userProfileService.CreateUser(userToCreate);
                 userAuthenticationModel.UserID = userCreated.UserId;
                 userAuthenticationModel.PartyID = userCreated.PartyId;
-             }
+            }
         }
 
         /// <summary>
@@ -971,7 +974,7 @@ namespace Altinn.Platform.Authentication.Controllers
             // it MAY use the http scheme, provided that the Client Type is confidential, as defined in Section 2.1 of OAuth 2.0, and
             // provided the OP allows the use of http Redirection URIs in this case. The Redirection URI MAY use an alternate scheme,
             // such as one that is intended to identify a callback into a native application.
-            if (!authorizationEndpoint.Contains("?"))
+            if (!authorizationEndpoint.Contains('?'))
             {
                 authorizationEndpoint += "?redirect_uri=" + redirect_uri;
             }
@@ -1029,7 +1032,7 @@ namespace Altinn.Platform.Authentication.Controllers
 
         private void CreateGoToCookie(HttpContext httpContext, string goToUrl)
         {
-            httpContext.Response.Cookies.Append(_generalSettings.AuthnGotToCookieName, goToUrl);
+            httpContext.Response.Cookies.Append(_generalSettings.AuthnGoToCookieName, goToUrl);
         }
 
         private async Task CreateTokenCookie(UserAuthenticationModel userAuthentication)
@@ -1072,7 +1075,7 @@ namespace Altinn.Platform.Authentication.Controllers
 
         private static string HashNonce(string nonce)
         {
-            using (SHA256 nonceHash = SHA256Managed.Create())
+            using (SHA256 nonceHash = SHA256.Create())
             {
                 byte[] byteArrayResultOfRawData = Encoding.UTF8.GetBytes(nonce);
                 byte[] byteArrayResult = nonceHash.ComputeHash(byteArrayResultOfRawData);
