@@ -7,6 +7,7 @@ using System.Security.Claims;
 using System.Text.Json;
 using System.Threading.Tasks;
 
+using Altinn.Platform.Authentication.Controllers;
 using Altinn.Platform.Authentication.Model;
 using Altinn.Platform.Authentication.Services;
 using Altinn.Platform.Authentication.Services.Interfaces;
@@ -27,19 +28,19 @@ using Xunit;
 
 namespace Altinn.Platform.Authentication.Tests.Controllers
 {
-    public class IntrospectionControllerTest : IClassFixture<WebApplicationFactory<Startup>>
+    public class IntrospectionControllerTest : IClassFixture<WebApplicationFactory<IntrospectionController>>
     {
-        private WebApplicationFactory<Startup> _factory;
+        private readonly WebApplicationFactory<IntrospectionController> _factory;
         private readonly Mock<IEFormidlingAccessValidator> _eformidlingValidatorService;
-        private string _baseUrl = "/authentication/api/v1/introspection";
-        private ClaimsPrincipal _testPrincipal;
+        private readonly string _baseUrl = "/authentication/api/v1/introspection";
+        private readonly ClaimsPrincipal _testPrincipal;
         private readonly JsonSerializerOptions _options;
 
         /// <summary>
         /// Initialises a new instance of the <see cref="IntrospectionControllerTest"/> class with the given WebApplicationFactory.
         /// </summary>
         /// <param name="factory">The WebApplicationFactory to use when creating a test server.</param>   
-        public IntrospectionControllerTest(WebApplicationFactory<Startup> factory)
+        public IntrospectionControllerTest(WebApplicationFactory<IntrospectionController> factory)
         {
             _factory = factory;
             _eformidlingValidatorService = new Mock<IEFormidlingAccessValidator>();
@@ -169,7 +170,6 @@ namespace Altinn.Platform.Authentication.Tests.Controllers
 
         private HttpClient GetTestClient(IEFormidlingAccessValidator eFormidlingAccessValidatorMock)
         {
-            Program.ConfigureSetupLogging();
             HttpClient client = _factory.WithWebHostBuilder(builder =>
             {
                 builder.ConfigureTestServices(services =>
